requirements.txt
# ABOUTME: Python dependencies for the comms AI portfolio project.
# ABOUTME: Includes Anthropic SDK for Claude-powered workflows.
anthropic>=0.42.0
python-dotenv>=1.0.0
pyyaml>=6.0
